requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-cosmos==4.9.0
azure-identity==1.20.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b35
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fixedint==0.1.6
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.1.1
h11==0.14.0
hatchling==1.27.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-dbapi==0.51b0
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-instrumentation-psycopg2==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-instrumentation-wsgi==0.51b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pluggy==1.5.0
portalocker==2.10.1
psutil==6.1.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.0.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
stripe==11.6.0
trove-classifiers==2025.3.3.18
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0
